+++ datafusion/core/src/physical_plan/mod.rs
@@ -552,7 +552,6 @@ pub mod explain;
 use crate::execution::context::TaskContext;
 pub use datafusion_physical_expr::expressions;
 
-pub mod aggregate_rule;
 pub mod file_format;
 pub mod filter;
 pub mod functions;
