+++ datafusion/physical-expr/src/aggregate/average.rs
@@ -21,6 +21,8 @@ use std::any::Any;
 use std::convert::TryFrom;
 use std::sync::Arc;
 
+use crate::aggregate::sum;
+use crate::expressions::format_state_name;
 use crate::{AggregateExpr, PhysicalExpr};
 use arrow::compute;
 use arrow::datatypes::DataType;
@@ -31,8 +33,6 @@ use arrow::{
 use datafusion_common::ScalarValue;
 use datafusion_common::{DataFusionError, Result};
 use datafusion_expr::Accumulator;
-
-use super::{format_state_name, sum};
 
 /// AVG aggregate expression
 #[derive(Debug)]
