+++ datafusion/physical-expr/src/aggregate/count.rs
@@ -31,7 +31,7 @@ use datafusion_common::Result;
 use datafusion_common::ScalarValue;
 use datafusion_expr::Accumulator;
 
-use super::format_state_name;
+use crate::expressions::format_state_name;
 
 /// COUNT aggregate expression
 /// Returns the amount of non-null values of the given expression.
