+++ datafusion/physical-expr/src/aggregate/min_max.rs
@@ -37,7 +37,7 @@ use datafusion_common::ScalarValue;
 use datafusion_common::{DataFusionError, Result};
 use datafusion_expr::Accumulator;
 
-use super::format_state_name;
+use crate::expressions::format_state_name;
 use arrow::array::Array;
 use arrow::array::DecimalArray;
