+++ datafusion/physical-expr/src/aggregate/stddev.rs
@@ -20,42 +20,37 @@
 use std::any::Any;
 use std::sync::Arc;
 
+use crate::aggregate::stats::StatsType;
+use crate::aggregate::variance::VarianceAccumulator;
+use crate::expressions::format_state_name;
 use crate::{AggregateExpr, PhysicalExpr};
-use arrow::array::Float64Array;
-use arrow::{
-    array::{ArrayRef, UInt64Array},
-    compute::cast,
-    datatypes::DataType,
-    datatypes::Field,
-};
+use arrow::{array::ArrayRef, datatypes::DataType, datatypes::Field};
 use datafusion_common::ScalarValue;
 use datafusion_common::{DataFusionError, Result};
 use datafusion_expr::Accumulator;
 
-use super::{format_state_name, StatsType};
-
-/// VAR and VAR_SAMP aggregate expression
+/// STDDEV and STDDEV_SAMP (standard deviation) aggregate expression
 #[derive(Debug)]
-pub struct Variance {
+pub struct Stddev {
     name: String,
     expr: Arc<dyn PhysicalExpr>,
 }
 
-/// VAR_POP aggregate expression
+/// STDDEV_POP population aggregate expression
 #[derive(Debug)]
-pub struct VariancePop {
+pub struct StddevPop {
     name: String,
     expr: Arc<dyn PhysicalExpr>,
 }
 
-impl Variance {
-    /// Create a new VARIANCE aggregate function
+impl Stddev {
+    /// Create a new STDDEV aggregate function
     pub fn new(
         expr: Arc<dyn PhysicalExpr>,
         name: impl Into<String>,
         data_type: DataType,
     ) -> Self {
-        // the result of variance just support FLOAT64 data type.
+        // the result of stddev just support FLOAT64 and Decimal data type.
         assert!(matches!(data_type, DataType::Float64));
         Self {
             name: name.into(),
@@ -64,7 +59,7 @@ impl Variance {
     }
 }
 
-impl AggregateExpr for Variance {
+impl AggregateExpr for Stddev {
     /// Return a reference to Any that can be used for downcasting
     fn as_any(&self) -> &dyn Any {
         self
@@ -75,7 +70,7 @@ impl AggregateExpr for Variance {
     }
 
     fn create_accumulator(&self) -> Result<Box<dyn Accumulator>> {
-        Ok(Box::new(VarianceAccumulator::try_new(StatsType::Sample)?))
+        Ok(Box::new(StddevAccumulator::try_new(StatsType::Sample)?))
     }
 
     fn state_fields(&self) -> Result<Vec<Field>> {
@@ -107,14 +102,14 @@ impl AggregateExpr for Variance {
     }
 }
 
-impl VariancePop {
-    /// Create a new VAR_POP aggregate function
+impl StddevPop {
+    /// Create a new STDDEV aggregate function
     pub fn new(
         expr: Arc<dyn PhysicalExpr>,
         name: impl Into<String>,
         data_type: DataType,
     ) -> Self {
-        // the result of variance just support FLOAT64 data type.
+        // the result of stddev just support FLOAT64 and Decimal data type.
         assert!(matches!(data_type, DataType::Float64));
         Self {
             name: name.into(),
@@ -123,7 +118,7 @@ impl VariancePop {
     }
 }
 
-impl AggregateExpr for VariancePop {
+impl AggregateExpr for StddevPop {
     /// Return a reference to Any that can be used for downcasting
     fn as_any(&self) -> &dyn Any {
         self
@@ -134,9 +129,7 @@ impl AggregateExpr for VariancePop {
     }
 
     fn create_accumulator(&self) -> Result<Box<dyn Accumulator>> {
-        Ok(Box::new(VarianceAccumulator::try_new(
-            StatsType::Population,
-        )?))
+        Ok(Box::new(StddevAccumulator::try_new(StatsType::Population)?))
     }
 
     fn state_fields(&self) -> Result<Vec<Field>> {
@@ -167,128 +160,55 @@ impl AggregateExpr for VariancePop {
         &self.name
     }
 }
-
-/// An accumulator to compute variance
-/// The algrithm used is an online implementation and numerically stable. It is based on this paper:
-/// Welford, B. P. (1962). "Note on a method for calculating corrected sums of squares and products".
-/// Technometrics. 4 (3): 419–420. doi:10.2307/1266577. JSTOR 1266577.
-///
-/// The algorithm has been analyzed here:
-/// Ling, Robert F. (1974). "Comparison of Several Algorithms for Computing Sample Means and Variances".
-/// Journal of the American Statistical Association. 69 (348): 859–866. doi:10.2307/2286154. JSTOR 2286154.
-
+/// An accumulator to compute the average
 #[derive(Debug)]
-pub struct VarianceAccumulator {
-    m2: f64,
-    mean: f64,
-    count: u64,
-    stats_type: StatsType,
+pub struct StddevAccumulator {
+    variance: VarianceAccumulator,
 }
 
-impl VarianceAccumulator {
-    /// Creates a new `VarianceAccumulator`
+impl StddevAccumulator {
+    /// Creates a new `StddevAccumulator`
     pub fn try_new(s_type: StatsType) -> Result<Self> {
         Ok(Self {
-            m2: 0_f64,
-            mean: 0_f64,
-            count: 0_u64,
-            stats_type: s_type,
+            variance: VarianceAccumulator::try_new(s_type)?,
         })
     }
 
-    pub fn get_count(&self) -> u64 {
-        self.count
-    }
-
-    pub fn get_mean(&self) -> f64 {
-        self.mean
-    }
-
     pub fn get_m2(&self) -> f64 {
-        self.m2
+        self.variance.get_m2()
     }
 }
 
-impl Accumulator for VarianceAccumulator {
+impl Accumulator for StddevAccumulator {
     fn state(&self) -> Result<Vec<ScalarValue>> {
         Ok(vec![
-            ScalarValue::from(self.count),
-            ScalarValue::from(self.mean),
-            ScalarValue::from(self.m2),
+            ScalarValue::from(self.variance.get_count()),
+            ScalarValue::from(self.variance.get_mean()),
+            ScalarValue::from(self.variance.get_m2()),
         ])
     }
 
     fn update_batch(&mut self, values: &[ArrayRef]) -> Result<()> {
-        let values = &cast(&values[0], &DataType::Float64)?;
-        let arr = values
-            .as_any()
-            .downcast_ref::<Float64Array>()
-            .unwrap()
-            .iter()
-            .flatten();
-
-        for value in arr {
-            let new_count = self.count + 1;
-            let delta1 = value - self.mean;
-            let new_mean = delta1 / new_count as f64 + self.mean;
-            let delta2 = value - new_mean;
-            let new_m2 = self.m2 + delta1 * delta2;
-
-            self.count += 1;
-            self.mean = new_mean;
-            self.m2 = new_m2;
-        }
-
-        Ok(())
+        self.variance.update_batch(values)
     }
 
     fn merge_batch(&mut self, states: &[ArrayRef]) -> Result<()> {
-        let counts = states[0].as_any().downcast_ref::<UInt64Array>().unwrap();
-        let means = states[1].as_any().downcast_ref::<Float64Array>().unwrap();
-        let m2s = states[2].as_any().downcast_ref::<Float64Array>().unwrap();
-
-        for i in 0..counts.len() {
-            let c = counts.value(i);
-            if c == 0_u64 {
-                continue;
-            }
-            let new_count = self.count + c;
-            let new_mean = self.mean * self.count as f64 / new_count as f64
-                + means.value(i) * c as f64 / new_count as f64;
-            let delta = self.mean - means.value(i);
-            let new_m2 = self.m2
-                + m2s.value(i)
-                + delta * delta * self.count as f64 * c as f64 / new_count as f64;
-
-            self.count = new_count;
-            self.mean = new_mean;
-            self.m2 = new_m2;
-        }
-        Ok(())
+        self.variance.merge_batch(states)
     }
 
     fn evaluate(&self) -> Result<ScalarValue> {
-        let count = match self.stats_type {
-            StatsType::Population => self.count,
-            StatsType::Sample => {
-                if self.count > 0 {
-                    self.count - 1
+        let variance = self.variance.evaluate()?;
+        match variance {
+            ScalarValue::Float64(e) => {
+                if e == None {
+                    Ok(ScalarValue::Float64(None))
                 } else {
-                    self.count
+                    Ok(ScalarValue::Float64(e.map(|f| f.sqrt())))
                 }
             }
-        };
-
-        if count <= 1 {
-            return Err(DataFusionError::Internal(
-                "At least two values are needed to calculate variance".to_string(),
-            ));
-        }
-
-        if self.count == 0 {
-            Ok(ScalarValue::Float64(None))
-        } else {
-            Ok(ScalarValue::Float64(Some(self.m2 / count as f64)))
+            _ => Err(DataFusionError::Internal(
+                "Variance should be f64".to_string(),
+            )),
         }
     }
 }
@@ -303,100 +223,99 @@ mod tests {
     use datafusion_common::Result;
 
     #[test]
-    fn variance_f64_1() -> Result<()> {
+    fn stddev_f64_1() -> Result<()> {
         let a: ArrayRef = Arc::new(Float64Array::from(vec![1_f64, 2_f64]));
         generic_test_op!(
             a,
             DataType::Float64,
-            VariancePop,
-            ScalarValue::from(0.25_f64),
+            StddevPop,
+            ScalarValue::from(0.5_f64),
             DataType::Float64
         )
     }
 
     #[test]
-    fn variance_f64_2() -> Result<()> {
-        let a: ArrayRef =
-            Arc::new(Float64Array::from(vec![1_f64, 2_f64, 3_f64, 4_f64, 5_f64]));
-        generic_test_op!(
-            a,
-            DataType::Float64,
-            VariancePop,
-            ScalarValue::from(2_f64),
-            DataType::Float64
-        )
-    }
-
-    #[test]
-    fn variance_f64_3() -> Result<()> {
-        let a: ArrayRef =
-            Arc::new(Float64Array::from(vec![1_f64, 2_f64, 3_f64, 4_f64, 5_f64]));
-        generic_test_op!(
-            a,
-            DataType::Float64,
-            Variance,
-            ScalarValue::from(2.5_f64),
-            DataType::Float64
-        )
-    }
-
-    #[test]
-    fn variance_f64_4() -> Result<()> {
+    fn stddev_f64_2() -> Result<()> {
         let a: ArrayRef = Arc::new(Float64Array::from(vec![1.1_f64, 2_f64, 3_f64]));
         generic_test_op!(
             a,
             DataType::Float64,
-            Variance,
-            ScalarValue::from(0.9033333333333333_f64),
+            StddevPop,
+            ScalarValue::from(0.7760297817881877),
             DataType::Float64
         )
     }
 
     #[test]
-    fn variance_i32() -> Result<()> {
+    fn stddev_f64_3() -> Result<()> {
+        let a: ArrayRef =
+            Arc::new(Float64Array::from(vec![1_f64, 2_f64, 3_f64, 4_f64, 5_f64]));
+        generic_test_op!(
+            a,
+            DataType::Float64,
+            StddevPop,
+            ScalarValue::from(std::f64::consts::SQRT_2),
+            DataType::Float64
+        )
+    }
+
+    #[test]
+    fn stddev_f64_4() -> Result<()> {
+        let a: ArrayRef = Arc::new(Float64Array::from(vec![1.1_f64, 2_f64, 3_f64]));
+        generic_test_op!(
+            a,
+            DataType::Float64,
+            Stddev,
+            ScalarValue::from(0.9504384952922168),
+            DataType::Float64
+        )
+    }
+
+    #[test]
+    fn stddev_i32() -> Result<()> {
         let a: ArrayRef = Arc::new(Int32Array::from(vec![1, 2, 3, 4, 5]));
         generic_test_op!(
             a,
             DataType::Int32,
-            VariancePop,
-            ScalarValue::from(2_f64),
+            StddevPop,
+            ScalarValue::from(std::f64::consts::SQRT_2),
             DataType::Float64
         )
     }
 
     #[test]
-    fn variance_u32() -> Result<()> {
+    fn stddev_u32() -> Result<()> {
         let a: ArrayRef =
             Arc::new(UInt32Array::from(vec![1_u32, 2_u32, 3_u32, 4_u32, 5_u32]));
         generic_test_op!(
             a,
             DataType::UInt32,
-            VariancePop,
-            ScalarValue::from(2.0f64),
+            StddevPop,
+            ScalarValue::from(std::f64::consts::SQRT_2),
             DataType::Float64
         )
     }
 
     #[test]
-    fn variance_f32() -> Result<()> {
+    fn stddev_f32() -> Result<()> {
         let a: ArrayRef =
             Arc::new(Float32Array::from(vec![1_f32, 2_f32, 3_f32, 4_f32, 5_f32]));
         generic_test_op!(
             a,
             DataType::Float32,
-            VariancePop,
-            ScalarValue::from(2_f64),
+            StddevPop,
+            ScalarValue::from(std::f64::consts::SQRT_2),
             DataType::Float64
         )
     }
 
     #[test]
-    fn test_variance_1_input() -> Result<()> {
+    fn test_stddev_1_input() -> Result<()> {
         let a: ArrayRef = Arc::new(Float64Array::from(vec![1_f64]));
         let schema = Schema::new(vec![Field::new("a", DataType::Float64, false)]);
         let batch = RecordBatch::try_new(Arc::new(schema.clone()), vec![a])?;
 
-        let agg = Arc::new(Variance::new(
+        let agg = Arc::new(Stddev::new(
             col("a", &schema)?,
             "bla".to_string(),
             DataType::Float64,
@@ -408,7 +327,7 @@ mod tests {
     }
 
     #[test]
-    fn variance_i32_with_nulls() -> Result<()> {
+    fn stddev_i32_with_nulls() -> Result<()> {
         let a: ArrayRef = Arc::new(Int32Array::from(vec![
             Some(1),
             None,
@@ -419,19 +338,19 @@ mod tests {
         generic_test_op!(
             a,
             DataType::Int32,
-            VariancePop,
-            ScalarValue::from(2.1875f64),
+            StddevPop,
+            ScalarValue::from(1.479019945774904),
             DataType::Float64
         )
     }
 
     #[test]
-    fn variance_i32_all_nulls() -> Result<()> {
+    fn stddev_i32_all_nulls() -> Result<()> {
         let a: ArrayRef = Arc::new(Int32Array::from(vec![None, None]));
         let schema = Schema::new(vec![Field::new("a", DataType::Int32, false)]);
         let batch = RecordBatch::try_new(Arc::new(schema.clone()), vec![a])?;
 
-        let agg = Arc::new(Variance::new(
+        let agg = Arc::new(Stddev::new(
             col("a", &schema)?,
             "bla".to_string(),
             DataType::Float64,
@@ -443,7 +362,7 @@ mod tests {
     }
 
     #[test]
-    fn variance_f64_merge_1() -> Result<()> {
+    fn stddev_f64_merge_1() -> Result<()> {
         let a = Arc::new(Float64Array::from(vec![1_f64, 2_f64, 3_f64]));
         let b = Arc::new(Float64Array::from(vec![4_f64, 5_f64]));
 
@@ -452,26 +371,26 @@ mod tests {
         let batch1 = RecordBatch::try_new(Arc::new(schema.clone()), vec![a])?;
         let batch2 = RecordBatch::try_new(Arc::new(schema.clone()), vec![b])?;
 
-        let agg1 = Arc::new(VariancePop::new(
+        let agg1 = Arc::new(StddevPop::new(
             col("a", &schema)?,
             "bla".to_string(),
             DataType::Float64,
         ));
 
-        let agg2 = Arc::new(VariancePop::new(
+        let agg2 = Arc::new(StddevPop::new(
             col("a", &schema)?,
             "bla".to_string(),
             DataType::Float64,
         ));
 
         let actual = merge(&batch1, &batch2, agg1, agg2)?;
-        assert!(actual == ScalarValue::from(2_f64));
+        assert!(actual == ScalarValue::from(std::f64::consts::SQRT_2));
 
         Ok(())
     }
 
     #[test]
-    fn variance_f64_merge_2() -> Result<()> {
+    fn stddev_f64_merge_2() -> Result<()> {
         let a = Arc::new(Float64Array::from(vec![1_f64, 2_f64, 3_f64, 4_f64, 5_f64]));
         let b = Arc::new(Float64Array::from(vec![None]));
 
@@ -480,20 +399,20 @@ mod tests {
         let batch1 = RecordBatch::try_new(Arc::new(schema.clone()), vec![a])?;
         let batch2 = RecordBatch::try_new(Arc::new(schema.clone()), vec![b])?;
 
-        let agg1 = Arc::new(VariancePop::new(
+        let agg1 = Arc::new(StddevPop::new(
             col("a", &schema)?,
             "bla".to_string(),
             DataType::Float64,
         ));
 
-        let agg2 = Arc::new(VariancePop::new(
+        let agg2 = Arc::new(StddevPop::new(
             col("a", &schema)?,
             "bla".to_string(),
             DataType::Float64,
         ));
 
         let actual = merge(&batch1, &batch2, agg1, agg2)?;
-        assert!(actual == ScalarValue::from(2_f64));
+        assert!(actual == ScalarValue::from(std::f64::consts::SQRT_2));
 
         Ok(())
     }
