+++ datafusion/physical-expr/src/window/mod.rs
@@ -18,7 +18,12 @@
 mod aggregate;
 mod built_in;
 mod built_in_window_function_expr;
+pub(crate) mod cume_dist;
+pub(crate) mod lead_lag;
+pub(crate) mod nth_value;
 pub(crate) mod partition_evaluator;
+pub(crate) mod rank;
+pub(crate) mod row_number;
 mod window_expr;
 
 pub use aggregate::AggregateWindowExpr;
